+++ modules/engine/test/exec-describe-routes-test.js
@@ -16,19 +16,18 @@
 
 var _ = require('underscore'),
     Engine = require('../lib/engine'),
-    util = require('util'),
     logger = require('winston');
 
 logger.remove(logger.transports.Console);
 logger.add(logger.transports.Console, {level: 'error'});
 
-exports["describe route '/foo/bar/{selector}?userid={userId}&itemid={itemId}' using method get"] = function (test) {
+exports['describe route "/foo/bar/{selector}?userid={userId}&itemid={itemId}" using method get'] = function (test) {
     var engine = new Engine({
             tables : __dirname + '/mock-routes/tables',
             routes : __dirname + '/mock-routes/routes',
             config : __dirname + '/config/dev.json'
         });
-    var q = "describe route '/foo/bar/{selector}?userid={userId}&itemid={itemId}' using method get";
+    var q = 'describe route "/foo/bar/{selector}?userid={userId}&itemid={itemId}" using method get';
         engine.exec(q, function(err, list) {
 
             if (err) {
@@ -50,13 +49,13 @@ exports["describe route '/foo/bar/{selector}?userid={userId}&itemid={itemId}' us
 
 };
 
-exports["describe route '/ping/pong' using method put"] = function (test) {
+exports['describe route "/ping/pong" using method put'] = function (test) {
     var engine = new Engine({
             tables : __dirname + '/mock-routes/tables',
             routes : __dirname + '/mock-routes/routes',
             config : __dirname + '/config/dev.json'
         });
-    var q = "describe route '/ping/pong' using method put";
+    var q = 'describe route "/ping/pong" using method put';
         engine.exec(q, function(err, list) {
 
             if (err) {
@@ -71,9 +70,31 @@ exports["describe route '/ping/pong' using method put"] = function (test) {
                 test.ok(_.isArray(list.body.tables), 'list.body.tables is not array');
                 test.ok(list.body.tables.length == 1 , 'Expected length = 1');
                 test.ok(_.isArray(list.body.related), 'list.body.related is not array');
-                test.ok(list.body.related.length == 1 , 'Expected length = 1');            }
+                test.ok(list.body.related.length == 1 , 'Expected length = 1');
+            }
             test.done();
+        }
+    );
+};
+
+// This method is not supported for the route
+exports['describe route "/ping/pong" using method patch'] = function (test) {
+    var engine = new Engine({
+            tables : __dirname + '/mock-routes/tables',
+            routes : __dirname + '/mock-routes/routes',
+            config : __dirname + '/config/dev.json'
         });
+    var q = 'describe route "/ping/pong" using method patch';
+        engine.exec(q, function(err, list) {
+            if (err) {
+                test.ok('Failed as expected');
+            }
+            else {
+                test.fail('Expected to fail');
+            }
+            test.done();
+        }
+    );
 };
 
 exports['check routes in desc table'] = function (test) {
@@ -99,10 +120,41 @@ exports['check routes in desc table'] = function (test) {
                 test.equals(list.headers['content-type'], 'application/json', 'JSON expected');
                 test.equals(list.body.name, 'testing.for.post');
                 test.equals(list.body.about, '/table?name=testing.for.post');
-                test.ok(list.body.select, "expected statement select");
-                test.ok(list.body.select.request, "expected request for statement select");
+                test.ok(list.body.select, 'expected statement select');
+                test.equals(list.body.select.method, 'post', 'expected method for statement select');
+                test.equals(list.body.select.uri, 'http://localhost:80126/ping/pong', 'expected uri for statement select');
                 test.ok(_.isArray(list.body.routes), 'expected list.body.routes to be array');
                 test.ok(list.body.routes.length > 0, 'expected list.body.routes to not be empty');
+                test.done();
+            }
+        }
+    };
+
+    engine.exec(opts);
+};
+
+exports['check table deps for route with insert'] = function (test) {
+    var engine = new Engine({
+        tables : __dirname + '/mock-routes/tables',
+        routes : __dirname + '/mock-routes/routes',
+        config : __dirname + '/config/dev.json'
+    });
+
+    var opts = {
+        request: {
+            headers: {},
+            params: {fromRoute: true}
+        },
+        script: 'describe route "/bitly/shorten" using method post',
+        cb: function(err, list) {
+            if (err) {
+                test.fail('got error: ' + err.stack);
+                test.done();
+            }
+            else {
+                test.equals(list.headers['content-type'], 'application/json', 'JSON expected');
+                test.equals(list.body.path, '/bitly/shorten');
+                test.equals(list.body.tables[0], '/table?name=bitly.shorten');
                 test.done();
             }
         }